pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Voice agent workspace"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Web framework and server
    "fastapi",
    "uvicorn[standard]",
    "websockets",
    # HTTP clients
    "httpx",
    "sse-starlette",
    # Data validation
    "pydantic",
    "pyyaml",
    # Audio processing
    "sounddevice",
    "numpy",
    "matplotlib",
    # ML/AI frameworks
    "torch==2.8.*",
    "torchaudio",
    "faster-whisper>=1.2.1",
    # OCR
    "pyautogui",
    "paddleocr",
    "paddlepaddle",
    "pillow",
    "rapidfuzz",
    "google-genai",
    # LLM providers
    "ollama",
    "llama-cpp-python",
    # TTS
    "edge-tts==7.2.1",
    "pydub",
    "scipy",  # For audio resampling (optional but recommended)
    # Hotkey management
    "pynput",
    # OBS
    "obs-websocket-py",
]


[tool.uv.sources]
# Install PyTorch with CUDA support on Linux/Windows (CUDA doesn't exist for Mac).
# This applies to all workspace members that depend on torch.
torch = [
    { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
    { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cuda"
# Use PyTorch built for NVIDIA Toolkit version 12.6.
# Available versions: https://pytorch.org/get-started/locally/
url = "https://download.pytorch.org/whl/cu126"
# Only use this index when explicitly requested by `tool.uv.sources`.
explicit = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
